+++ Source/Scene/ImageryProvider.js
@@ -346,11 +346,13 @@ define([
         } else if (defined(imageryProvider.tileDiscardPolicy)) {
             return resource.fetchImage({
                 preferBlob : true,
+                preferImageBitmap : true,
                 flipY : true
             });
         }
 
         return resource.fetchImage({
+            preferImageBitmap : true,
             flipY : true
         });
     };