+++ Source/Scene/Material.js
@@ -817,9 +817,7 @@ define([
                     } else if (crnRegex.test(uniformValue)) {
                         promise = loadCRN(resource);
                     } else {
-                        promise = resource.fetchImage({
-                            flipY : true
-                        });
+                        promise = resource.fetchImage();
                     }
                     when(promise, function(image) {
                         material._loadedImages.push({
@@ -868,17 +866,13 @@ define([
                 uniformValue.positiveZ + uniformValue.negativeZ;
 
             if (path !== material._texturePaths[uniformId]) {
-                var fetchOptions = {
-                    flipY : true
-                };
-
                 var promises = [
-                    Resource.createIfNeeded(uniformValue.positiveX).fetchImage(fetchOptions),
-                    Resource.createIfNeeded(uniformValue.negativeX).fetchImage(fetchOptions),
-                    Resource.createIfNeeded(uniformValue.positiveY).fetchImage(fetchOptions),
-                    Resource.createIfNeeded(uniformValue.negativeY).fetchImage(fetchOptions),
-                    Resource.createIfNeeded(uniformValue.positiveZ).fetchImage(fetchOptions),
-                    Resource.createIfNeeded(uniformValue.negativeZ).fetchImage(fetchOptions)
+                    Resource.createIfNeeded(uniformValue.positiveX).fetchImage(),
+                    Resource.createIfNeeded(uniformValue.negativeX).fetchImage(),
+                    Resource.createIfNeeded(uniformValue.positiveY).fetchImage(),
+                    Resource.createIfNeeded(uniformValue.negativeY).fetchImage(),
+                    Resource.createIfNeeded(uniformValue.positiveZ).fetchImage(),
+                    Resource.createIfNeeded(uniformValue.negativeZ).fetchImage()
                 ];
 
                 when.all(promises).then(function(images) {