+++ Source/Scene/PostProcessStage.js
@@ -671,9 +671,7 @@ define([
                     url : stageNameUrlOrImage
                 });
 
-                promises.push(resource.fetchImage({
-                    flipY : true
-                }).then(createLoadImageFunction(stage, name)));
+                promises.push(resource.fetchImage().then(createLoadImageFunction(stage, name)));
             } else {
                 stage._texturesToCreate.push({
                     name : name,